refactor: Migrate battery helper feature from core package to a dedicated screen.
