+++ src/screens/battery-helper/components/battery-status-card.tsx
@@ -1,11 +1,11 @@
 import { Ionicons } from '@expo/vector-icons';
 import { View } from 'react-native';
-import { getThemeColors } from '../../constants/Colors';
-import { useTheme } from '../../contexts/ThemeContext';
-import { getBatteryColor } from '../../helpers/battery_helper';
-import { batteryStatusCardStyles } from '../../styles/battery_status_card.styles';
-import { ThemedText } from '../ThemedText';
-import { ThemedView } from '../ThemedView';
+import { ThemedText } from '../../../shared/components/ThemedText';
+import { ThemedView } from '../../../shared/components/ThemedView';
+import { getThemeColors } from '../../../shared/constants/Colors';
+import { useTheme } from '../../../shared/contexts/theme-context';
+import { getBatteryColor } from '../../../shared/helpers/battery_helper';
+import { batteryStatusCardStyles } from '../styles/battery-status-card.styles';
 
 interface BatteryStatusCardProps {
   batteryLevel: number;
@@ -120,4 +120,3 @@ export function BatteryStatusCard({
     </ThemedView>
   );
 }
-
