+++ src/screens/battery-helper/components/device-info-card.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from 'react';
 import { ActivityIndicator, View } from 'react-native';
-import { getThemeColors } from '../../constants/Colors';
-import { useTheme } from '../../contexts/ThemeContext';
-import { DeviceInfo as DeviceInfoHelper } from '../../helpers/device-info';
-import { deviceInfoCardStyles } from '../../styles/device_info_card.styles';
-import { ThemedText } from '../ThemedText';
-import { ThemedView } from '../ThemedView';
+import { ThemedText } from '../../../shared/components/ThemedText';
+import { ThemedView } from '../../../shared/components/ThemedView';
+import { getThemeColors } from '../../../shared/constants/Colors';
+import { useTheme } from '../../../shared/contexts/theme-context';
+import { DeviceInfo as DeviceInfoHelper } from '../../../shared/helpers/device-info';
+import { deviceInfoCardStyles } from '../styles/device-info-card.styles';
 
 interface DeviceInfoCardProps {
   getDeviceInformation: () => Promise<DeviceInfoHelper | null>;
@@ -48,7 +48,7 @@ export function DeviceInfoCard({ getDeviceInformation }: DeviceInfoCardProps) {
             Device Information
           </ThemedText>
         </View>
-        <ActivityIndicator size="small" color={colors.primary} />
+        <ActivityIndicator size="small" color={colors.tint} />
       </ThemedView>
     );
   }
