+++ src/screens/battery-helper/components/low-power-mode-card.tsx
@@ -1,9 +1,9 @@
 import { Button, Platform, View } from 'react-native';
-import { getThemeColors } from '../../constants/Colors';
-import { useTheme } from '../../contexts/ThemeContext';
-import { lowPowerModeCardStyles } from '../../styles/low_power_mode_card.styles';
-import { ThemedText } from '../ThemedText';
-import { ThemedView } from '../ThemedView';
+import { ThemedText } from '../../../shared/components/ThemedText';
+import { ThemedView } from '../../../shared/components/ThemedView';
+import { getThemeColors } from '../../../shared/constants/Colors';
+import { useTheme } from '../../../shared/contexts/theme-context';
+import { lowPowerModeCardStyles } from '../styles/low-power-mode-card.styles';
 
 interface LowPowerModeCardProps {
   lowPowerMode: boolean | null;
@@ -50,8 +50,8 @@ export function LowPowerModeCard({
                 lowPowerMode === true
                   ? colors.errorColor
                   : lowPowerMode === false
-                  ? colors.successColor
-                  : colors.surfaceBorder,
+                    ? colors.successColor
+                    : colors.surfaceBorder,
             },
           ]}
         />
@@ -74,7 +74,7 @@ export function LowPowerModeCard({
         <Button
           title="Open Settings"
           onPress={onOpenSettings}
-          color={colors.primary}
+          color={colors.tint}
         />
       )}
     </ThemedView>
