+++ src/screens/battery-helper/index.ts
@@ -1,0 +1,1 @@
+export * from './components/battery-helper-screen';
