+++ src/screens/battery-helper/styles/battery-helper-screen.styles.ts
