+++ src/screens/battery-helper/styles/battery-status-card.styles.ts
