+++ src/screens/battery-helper/styles/device-info-card.styles.ts
