+++ src/screens/battery-helper/styles/low-power-mode-card.styles.ts
