+++ src/shared/helpers/battery_helper.ts
