+++ src/shared/types/battery.ts
